AMAZONS/requirements.txt
# Core Dependencies
playwright>=1.40.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.2
fake-useragent>=1.4.0

# Optional but recommended
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
